rustc_target: spec: Ensure that a valid llvm_abiname value is set for PowerPC64(LE)
PowerPC64 ELF targets (effectively anything that isn't AIX) use either
the ELFv1 or ELFv2 ABI. The ELFv1 ABI is only specified for big endian
targets, while ELFv2 can be used by both little- and big-endian targets.
Make sure that, if an LLVM ABI is set, it is set to one of the two. AIX
does not set an LLVM ABI name, so it does not need to be accounted for
in any way other than allowing an unset value.

Signed-off-by: Jens Reidel <adrian@travitia.xyz>